Spent time trying to streamline build process and include git pulls to the image, with specific commit checkouts to avoid troublesome commits
Next work will be to introduce conf editing,
networking with external mongo and redis containers,
seeding the mongodb with indices (!!!), and moving the git
folders and database to external volumes (shrink image).

Final work will be to package a script that can call for
rebuids on demand, allows for changing versions and commits,
and passes correct proxying info if needed.